+++ plugins/plugin-quick-start-docs/redhat-plugin-quick-start-docs-ide/src/main/java/com/codenvy/redhat/plugin/quick/start/ide/QuickStartServiceClientImpl.java
