+++ plugins/plugin-quick-start-docs/redhat-plugin-quick-start-docs-ide/src/main/resources/com/codenvy/redhat/plugin/quick/start/ide/guide.css
