+++ plugins/plugin-quick-start-docs/redhat-plugin-quick-start-docs-server/src/main/java/com/codenvy/redhat/plugin/quick/start/server/QuickStartService.java
