+++ plugins/plugin-quick-start-docs/redhat-plugin-quick-start-docs-shared/src/main/java/com/codenvy/redhat/plugin/quick/start/shared/Constants.java
